backend/celery_worker.py
from celery import Celery
import os
from dotenv import load_dotenv

load_dotenv()

celery_app = Celery(
    "focusflow",
    broker=os.getenv("REDIS_URL", "redis://localhost:6379"),
    backend=os.getenv("REDIS_URL", "redis://localhost:6379"),
)

celery_app.conf.update(
    task_serializer="json",
    result_serializer="json",
    accept_content=["json"],
    timezone="UTC",
    enable_utc=True,
)
