backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
websockets==12.0
python-multipart==0.0.9
httpx==0.27.0

# ML core
torch==2.2.2
transformers==4.40.1
sentence-transformers==2.7.0
bertopic==0.16.2
xgboost==2.0.3
scikit-learn==1.4.2
numpy==1.26.4
faiss-cpu==1.8.0
datasets
evaluate

# STT
faster-whisper

# LLM
ollama

# Storage & Queue
redis==5.0.4
celery==5.4.0

# Utilities
slowapi
sentry-sdk
python-dotenv==1.0.1
pydantic==2.7.1
pydantic-settings==2.2.1
